ft39 Adding more commands to master

See merge request compsci308_2020spring/parser_parser_team09!16

+++ src/slogo/Model/Commands/TurtleCommands/ForwardCommand.java
@@ -46,8 +46,8 @@ public class ForwardCommand extends Command {
   }
 
   @Override
-  public Integer returnArgValue() {
-    return (int) this.returnArgValue;
+  public Double returnArgValue() {
+    return this.returnArgValue;
   }
 
 }
